refactored: new trashModal.js
