+++ src/main_page/keep_or_delete.html
@@ -154,6 +154,7 @@
 <script type="module" src="../settings.js"></script>
 <script type="module" src="../rename.js"></script>
 <script type="module" src="../sort.js"></script>
+<script type="module" src="../trashModal.js"></script>
 <script type="module" src="../currentIndex.js"></script>
 <script type="module" src="./keep_or_delete.js"></script>
